requirements.txt
asgiref==3.6.0
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudinary==1.33.0
cloudscraper==1.2.71
colorama==0.4.6
comtypes==1.1.14
cryptography==41.0.5
deckar01-ratelimit==3.0.2
decorator==4.4.2
deepl==1.15.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==4.1.4
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
filelock==3.10.7
fontawesome==5.10.1.post1
futures==3.0.5
goslate==1.5.4
gunicorn==21.2.0
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.7
lxml==4.9.3
moviepy==1.0.3
numpy==1.24.1
oauthlib==3.2.2
packaging==23.2
Pillow==9.3.0
pipenv==2023.3.20
platformdirs==3.2.0
proglog==0.1.10
psycopg2==2.9.9
pycparser==2.21
PyDictionary==2.0.1
PyJWT==2.8.0
pyparsing==3.1.1
pypiwin32==223
pyspellchecker==0.7.2
python3-openid==3.2.0
pyttsx3==2.90
pywin32==305
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
tk==0.1.0
tqdm==4.64.1
typing_extensions==4.9.0
tzdata==2022.7
urllib3==2.0.6
virtualenv==20.21.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
whitenoise==6.6.0
wordhoard==1.5.4
